Add basic create group gql mutation test
Also refactor individual mutation & query test to extract setup code
