+++ individual/tests/__init__.py
@@ -2,4 +2,5 @@ from .individual_service_test import IndividualServiceTest
 from .group_service_test import GroupServiceTest
 from .group_individual_service_test import GroupIndividualServiceTest
 from .graphql_query_test import IndividualGQLQueryTest
-from .graphql_mutation_test import IndividualGQLMutationTest
+from .graphql_mutation_individual_test import IndividualGQLMutationTest
+from .graphql_mutation_group_test import GroupGQLMutationTest
